+++ lib/generators/godmin/resource/templates/resource_controller.rb
@@ -4,14 +4,6 @@ require_dependency "<%= File.join(namespaced_path, "application_controller") %>"
 <% end -%>
 <% module_namespacing do -%>
 class <%= class_name.pluralize %>Controller < ApplicationController
-  include Godmin::Resource
-
-  def attrs_for_index
-    <%= @attributes.map(&:to_sym) %>
-  end
-
-  def attrs_for_form
-    <%= @attributes.map(&:to_sym) %>
-  end
+  include Godmin::Resources::ResourceController
 end
 <% end -%>