+++ lib/godmin/helpers/batch_actions.rb
@@ -5,8 +5,11 @@ module Godmin
         return unless include_batch_action_link?(options)
 
         link_to(
-          translate_scoped("batch_actions.labels.#{name}", default: name.to_s.titleize), "#",
-          class: "btn btn-default hidden", data: {
+          translate_scoped("batch_actions.labels.#{name}", default: name.to_s.titleize),
+          @resource_class,
+          method: :patch,
+          class: "btn btn-default hidden",
+          data: {
             behavior: "batch-actions-action-link",
             confirm: options[:confirm] ? translate_scoped("batch_actions.confirm_message") : false,
             value: name
@@ -18,8 +21,8 @@ module Godmin
 
       def include_batch_action_link?(options)
         (options[:only].nil? && options[:except].nil?) ||
-          (options[:only] && options[:only].include?(params[:scope].to_sym)) ||
-          (options[:except] && !options[:except].include?(params[:scope].to_sym))
+          (options[:only] && options[:only].include?(@resource_service.scope.to_sym)) ||
+          (options[:except] && !options[:except].include?(@resource_service.scope.to_sym))
       end
     end
   end